# src/FacilityManagement.Api/swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "FacilityManagement",
    "description": "",
    "termsOfService": "https://example.com/terms",
    "contact": {
      "name": "",
      "email": ""
    },
    "license": {
      "name": "Use under MIT",
      "url": "https://opensource.org/licenses/MIT"
    },
    "version": "v1"
  },
  "paths": {
...
    "/api/Profile/{profileId}": {
      "get": {
        "tags": [
          "Profile"
        ],
        "summary": "Get Profile by id.",
        "description": "Get Profile by id.",
        "operationId": "getProfileById",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "profileId",
            "required": true,
            "type": "string",
            "format": "uuid"
          }
        ],
        "responses": {
          "404": {
            "description": "Not Found",
            "schema": {
              "type": "string"
            }
          },
          "500": {
            "description": "Server Error"
          },
          "400": {
            "description": "Bad Request",
            "schema": {
              "$ref": "#/definitions/ProblemDetails"
            }
          },
          "200": {
            "description": "Success",
            "schema": {
              "$ref": "#/definitions/GetProfileByIdResponse"
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Profile"
        ],
        "summary": "Delete Profile.",
        "description": "Delete Profile.",
        "operationId": "removeProfile",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "profileId",
            "required": true,
            "type": "string",
            "format": "uuid"
          }
        ],
        "responses": {
          "500": {
            "description": "Server Error"
          },
          "400": {
            "description": "Bad Request",
            "schema": {
              "$ref": "#/definitions/ProblemDetails"
            }
          },
          "200": {
            "description": "Success",
            "schema": {
              "$ref": "#/definitions/RemoveProfileResponse"
            }
          }
        }
      }
    },
    "/api/Profile": {
      "get": {
        "tags": [
          "Profile"
        ],
        "summary": "Get Profiles.",
        "description": "Get Profiles.",
        "operationId": "getProfiles",
        "produces": [
          "application/json"
        ],
        "responses": {
          "500": {
            "description": "Server Error"
          },
          "400": {
            "description": "Bad Request",
            "schema": {
              "$ref": "#/definitions/ProblemDetails"
            }
          },
          "200": {
            "description": "Success",
            "schema": {
              "$ref": "#/definitions/GetProfilesResponse"
            }
          }
        }
      },
      "post": {
        "tags": [
          "Profile"
        ],
        "summary": "Create Profile.",
        "description": "Create Profile.",
        "operationId": "createProfile",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "schema": {
              "$ref": "#/definitions/CreateProfileRequest"
            }
          }
        ],
        "responses": {
          "500": {
            "description": "Server Error"
          },
          "400": {
            "description": "Bad Request",
            "schema": {
              "$ref": "#/definitions/ProblemDetails"
            }
          },
          "200": {
            "description": "Success",
            "schema": {
              "$ref": "#/definitions/CreateProfileResponse"
            }
          }
        }
      },
      "put": {
        "tags": [
          "Profile"
        ],
        "summary": "Update Profile.",
        "description": "Update Profile.",
        "operationId": "updateProfile",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "schema": {
              "$ref": "#/definitions/UpdateProfileRequest"
            }
          }
        ],
        "responses": {
          "500": {
            "description": "Server Error"
          },
          "400": {
            "description": "Bad Request",
            "schema": {
              "$ref": "#/definitions/ProblemDetails"
            }
          },
          "200": {
            "description": "Success",
            "schema": {
              "$ref": "#/definitions/UpdateProfileResponse"
            }
          }
        }
      }
    },
    "/api/Profile/page/{pageSize}/{index}": {
      "get": {
        "tags": [
          "Profile"
        ],
        "summary": "Get Profile Page.",
        "description": "Get Profile Page.",
        "operationId": "getProfilesPage",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "pageSize",
            "required": true,
            "type": "integer",
            "format": "int32"
          },
          {
            "in": "path",
            "name": "index",
            "required": true,
            "type": "integer",
            "format": "int32"
          }
        ],
        "responses": {
          "500": {
            "description": "Server Error"
          },
          "400": {
            "description": "Bad Request",
            "schema": {
              "$ref": "#/definitions/ProblemDetails"
            }
          },
          "200": {
            "description": "Success",
            "schema": {
              "$ref": "#/definitions/GetProfilesPageResponse"
            }
          }
        }
      }
    }
  },
  "definitions": {
...
    "CreateProfileRequest": {
      "type": "object",
      "properties": {
        "profile": {
          "$ref": "#/definitions/ProfileDto"
        }
      }
    },
    "CreateProfileResponse": {
      "type": "object",
      "properties": {
        "profile": {
          "$ref": "#/definitions/ProfileDto"
        },
        "validationErrors": {
          "type": "array",
          "items": {
            "type": "string"
          }
        }
      }
    },
...
    "GetProfileByIdResponse": {
      "type": "object",
      "properties": {
        "profile": {
          "$ref": "#/definitions/ProfileDto"
        },
        "validationErrors": {
          "type": "array",
          "items": {
            "type": "string"
          }
        }
      }
    },
    "GetProfilesPageResponse": {
      "type": "object",
      "properties": {
        "length": {
          "format": "int32",
          "type": "integer"
        },
        "entities": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/ProfileDto"
          }
        },
        "validationErrors": {
          "type": "array",
          "items": {
            "type": "string"
          }
        }
      }
    },
    "GetProfilesResponse": {
      "type": "object",
      "properties": {
        "profiles": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/ProfileDto"
          }
        },
        "validationErrors": {
          "type": "array",
          "items": {
            "type": "string"
          }
        }
      }
    },
...
    "ProblemDetails": {
      "type": "object",
      "properties": {
        "type": {
          "type": "string"
        },
        "title": {
          "type": "string"
        },
        "status": {
          "format": "int32",
          "type": "integer"
        },
        "detail": {
          "type": "string"
        },
        "instance": {
          "type": "string"
        }
      },
      "additionalProperties": { }
    },
    "ProfileDto": {
      "type": "object",
      "properties": {
        "profileId": {
          "format": "uuid",
          "type": "string"
        },
        "firstName": {
          "type": "string"
        },
        "lastName": {
          "type": "string"
        }
      }
    },
...
    "RemoveProfileResponse": {
      "type": "object",
      "properties": {
        "profile": {
          "$ref": "#/definitions/ProfileDto"
        },
        "validationErrors": {
          "type": "array",
          "items": {
            "type": "string"
          }
        }
      }
    },
...
    "UpdateProfileRequest": {
      "type": "object",
      "properties": {
        "profile": {
          "$ref": "#/definitions/ProfileDto"
        }
      }
    },
    "UpdateProfileResponse": {
      "type": "object",
      "properties": {
        "profile": {
          "$ref": "#/definitions/ProfileDto"
        },
        "validationErrors": {
          "type": "array",
          "items": {
            "type": "string"
          }
        }
      }
    }
  }
}